fix(cockpit): Merge rbac-utils.ts in rbac-schema.ts
Da die Schema-Definitionen nun nicht mehr generiert werden, können die Schema-Definitionen aus rbac-utils in rbac-shema übernommen werden.
